refactor(evm,facilitator): dedupe shared helpers, fix README
Addresses CodeRabbit review on #53.

- Move the three calldata-builder dummy meta-tx constants
  (`DUMMY_CHAIN_ID`, `DUMMY_METATX_HANDLER_ADDRESS`, `DUMMY_NONCE`)
  shared by `actions/create-offer-and-commit.ts` and
  `actions/create-offer-commit-and-redeem.ts` into
  `internal/metatx-calldata-constants.ts`.
- Move the ECDSA signature-packing helpers (`packEcdsaSignature` and
  its private helpers `assert32ByteHex` / `normalizeRecoveryId` /
  `WORD32_RE`) shared by `envelope/execute-meta-transaction.ts` and
  `envelope/execute-meta-transaction-with-token-auth.ts` into
  `internal/signature-helpers.ts`.
- Fix the facilitator README's opening status sentence — previously
  claimed the package was functional only for `tokenAuthStrategy: "none"`,
  which contradicted the following block that lists all four strategies
  as supported in `verify` and `settle`.

Both internal modules stay package-private (not re-exported from the
public surface). Behaviour and error messages are unchanged.
