+++ typescript/packages/evm/src/envelope/execute-meta-transaction-with-token-auth.ts
@@ -27,8 +27,9 @@
 // `TransferAuthorization`.
 
 import { erc20, metaTx } from "@bosonprotocol/core-sdk";
-import { concat, type Hex } from "viem";
+import type { Hex } from "viem";
 
+import { packEcdsaSignature } from "../internal/signature-helpers.js";
 import type { TxRequest } from "../types.js";
 import type { BuildExecuteMetaTransactionArgs } from "./execute-meta-transaction.js";
 
@@ -76,36 +77,4 @@ export function buildExecuteMetaTransactionWithTokenAuthTx(
     ],
   ) as Hex;
   return { to: args.escrowAddress, data };
-}
-
-/**
- * Pack a split ECDSA signature into the 65-byte `r ++ s ++ v` form the
- * contract's `LibSignature.recover` slices. Same shape as the sibling
- * `execute-meta-transaction.ts` packer; duplicated locally because the
- * `executeMetaTransaction` module keeps the helper private to avoid
- * leaking a low-level primitive into the package's public exports.
- */
-function packEcdsaSignature(sig: { r: Hex; s: Hex; v: number | bigint }): Hex {
-  assert32ByteHex(sig.r, "r");
-  assert32ByteHex(sig.s, "s");
-  const v = normalizeRecoveryId(sig.v);
-  if (v !== 27 && v !== 28) {
-    throw new Error(`@bosonprotocol/x402-evm: meta-tx signature v must be 27 or 28, got ${sig.v}`);
-  }
-  const vHex = `0x${v.toString(16).padStart(2, "0")}` as Hex;
-  return concat([sig.r, sig.s, vHex]);
-}
-
-const WORD32_RE = /^0x[0-9a-fA-F]{64}$/;
-
-function normalizeRecoveryId(v: number | bigint): number {
-  return typeof v === "bigint" ? Number(v) : v;
-}
-
-function assert32ByteHex(value: Hex, field: "r" | "s"): void {
-  if (!WORD32_RE.test(value)) {
-    throw new Error(
-      `@bosonprotocol/x402-evm: meta-tx signature ${field} must be a 32-byte hex value (0x-prefixed, 64 hex chars), got ${value}`,
-    );
-  }
 }
